Add db server handling

+++ deploy.sh
@@ -4,16 +4,22 @@ HOSTS=("$@")
 USERNAME=$USER
 NOTIFIER=`dirname $0`/notify.sh
 
+DB_SERVER='app3143'
+
 set -x
 for HOST in ${HOSTS[@]}; do
+    if [ $HOST == 'app3143' ]; then
+        echo "${HOST} is DB server. skip..."
+        continue
+    fi
     $NOTIFIER "$USERNAME -> $HOST: deploy start "
-    ssh -A "isucon@${HOST}" '
-        cd /home/isucon/isubata/webapp/ruby && git pull origin master ;\
-        sudo systemctl stop isubata.python.service;
-        sudo systemctl restart isubata.ruby.service;
-        sudo systemctl restart nginx;
-        # sudo service restart mysql;
-#       sudo service restart memcached;
+   ssh -A "isucon@${HOST}" '
+       cd /home/isucon/isubata/webapp/ruby && git pull origin master ;\
+       sudo systemctl stop isubata.python.service;
+       sudo systemctl restart isubata.ruby.service;
+       sudo systemctl restart nginx;
 '
-    $NOTIFIER "$USERNAME -> $HOST: deploy finish "
+#    $NOTIFIER "$USERNAME -> $HOST: deploy finish "
 done
+
+ssh isucon@$DB_SERVER sudo systemctl restart mysql
